山手線テーマの下部共通化と下部終端丸め処理
